requirements.txt
langchain
openai
pypdf
chromadb
tiktoken
python-dotenv